+++ crates/tests/src/lib.rs
@@ -1,6 +1,8 @@
 //! End to end tests for madara.
+#![cfg(test)]
 
 mod rpc;
+mod storage_proof;
 
 use anyhow::bail;
 use rstest::rstest;