+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- feat(v0.8.0-rc0): storage proofs for rpc version v0.8.0
 - feat(warp): added warp update to madara
 - docs(readme): updated README.md docs and added Docker Compose support
 - fix(log): define RUST_LOG=info by default